Adding button to create sprout
