+++ sprout/sprout_ui/src/static/css/create-sprout-button.css
@@ -1,0 +1,4 @@
+.btn btn-dark {
+    width: auto;
+    height: auto;
+}
