+++ sprout/sprout_ui/src/static/js/utilities/left_panel.js
@@ -2,6 +2,7 @@
 import { useNavigate } from 'react-router-dom';
 import '../../css/left-panel.css';
 import React from 'react';
+import CreateSproutButton from '../buttons/create_sprout_button';
 
 const LeftPanel = () => {
 
@@ -15,6 +16,7 @@ const LeftPanel = () => {
   return (
     <aside className="col-md-3 col-lg-2 d-md-block bg-light sidebar">
       <div className="sidebar-sticky">
+        <CreateSproutButton/>
         <ul className="nav flex-column">
           <li className="nav-item">
             <a className="nav-link active" onClick={handleHomeLink}>
